Add compiled main.js and update README
The main.js is taken from the original plugin's gh-page branch
(https://github.com/staticvariablejames/InsugarTrading), updated so that
data is retrieved from my own github.io page.
Given that this is a fork of a repository that seems to have updated
data recently, I hope the old main.js will work with the new data.

The README is also updated.